Area/backendNew/requirements.txt
Django==4.2.5
mysqlclient==2.2.0
django-bcrypt==0.9.2
django-cors-headers==4.2.0
requests==2.27.1
python-dotenv==0.15.0